+++ packages/solid/src/index.tsx
@@ -39,6 +39,7 @@ export type { LinkDirectiveOptions } from "./directives/link";
 export type {
   RouteViewProps,
   RouteViewMatchProps,
+  RouteViewSelfProps,
   RouteViewNotFoundProps,
 } from "./components/RouteView";